source_code/pyproject.toml
[tool.poetry]
name = "athena"
version = "0.1.0"
description = "A Django project for assisting with BOJ (Baekjoon Online Judge) problems"
authors = ["shinkeonkim <singun11@kookmin.ac.kr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
markdown = "^3.7"
django-environ = "^0.12.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
django-silk = "^5.3.2"
django-extensions = "^3.2.3"
django-postgres-extra = "^2.0.8"
uvicorn = "^0.34.0"
lorem = "^0.1.1"
colorlog = "^6.9.0"
aiohttp = "^3.11.18"
beautifulsoup4 = "^4.13.4"
celery = "^5.5.2"
django-celery-beat = "^2.8.0"
django-celery-results = "^2.6.0"
redis = "^6.0.0"
flower = "^2.0.1"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
drf-extensions = "^0.8.0"
django-guid = "^3.5.1"
drf-api-logger = "^1.1.19"
drf-writable-nested = "^0.7.2"
django-markdownx = "^4.0.9"
nplusone = "^1.0.0"
django-filtering = "^0.0.7"
django-filter = "^25.1"
django-admin-autocomplete-filter = "^0.7.1"
django-unfold = "^0.57.0"
django-import-export = "^4.3.7"
django-guardian = "^3.0.0"
django-simple-history = "^3.8.0"
lxml = "^5.2.0"
duckduckgo-search = "^8.0.1"
sentence-transformers = "^2.5.1"
numpy = "^1.26.4"
pydantic = "^2.6.3"
openai = "^1.82.0"
grpcio = "^1.72.1"
grpcio-tools = "^1.72.1"
protobuf = "^6.31.1"


[tool.poetry.dev-dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.9.3"
mypy = "^1.8.0"
autoflake = "^2.2.0"
coverage = "^7.7.1"
pytest = "^8.3.5"
pytest-django = "^4.10.0"
pytest-cov = "^6.0.0"
factory-boy = "^3.3.3"
faker = "^35.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
